+++ changelog.md
@@ -1,6 +1,8 @@
 # ObjectIDDisplay Changelog
 ## v1.0.5
 - Fix a bug where the mod wouldn't work without RaZoom's Object Groups mod. Not sure how that happened.
+- Redo positioning for object ID displays to always be relative to the object button.
+- Removed `Extra Safety Mode` setting.
 ## v1.0.4
 - Attempt to clean up the code.
 - Attempt to optimize the code, with help from Alphalaneous.